+++ dvd_cell.h
@@ -8,15 +8,15 @@ struct dvd_cell {
 	uint8_t cell;
 	char length[DVD_CELL_LENGTH + 1];
 	uint32_t msecs;
-	uint32_t first_sector;
-	uint32_t last_sector;
-	ssize_t blocks;
-	ssize_t filesize;
+	uint64_t first_sector;
+	uint64_t last_sector;
+	uint64_t blocks;
+	uint64_t filesize;
 };
 
-uint32_t dvd_cell_first_sector(const ifo_handle_t *vmg_ifo, const ifo_handle_t *vts_ifo, const uint16_t track_number, const uint8_t cell_number);
+uint64_t dvd_cell_first_sector(const ifo_handle_t *vmg_ifo, const ifo_handle_t *vts_ifo, const uint16_t track_number, const uint8_t cell_number);
 
-uint32_t dvd_cell_last_sector(const ifo_handle_t *vmg_ifo, const ifo_handle_t *vts_ifo, const uint16_t track_number, const uint8_t cell_number);
+uint64_t dvd_cell_last_sector(const ifo_handle_t *vmg_ifo, const ifo_handle_t *vts_ifo, const uint16_t track_number, const uint8_t cell_number);
 
 /**
  * Helper function to get the total number of blocks from a cell.
@@ -29,9 +29,9 @@ uint32_t dvd_cell_last_sector(const ifo_handle_t *vmg_ifo, const ifo_handle_t *v
  * naming.
  *
  */
-ssize_t dvd_cell_blocks(const ifo_handle_t *vmg_ifo, const ifo_handle_t *vts_ifo, const uint16_t track_number, const uint8_t cell_number);
+uint64_t dvd_cell_blocks(const ifo_handle_t *vmg_ifo, const ifo_handle_t *vts_ifo, const uint16_t track_number, const uint8_t cell_number);
 
-ssize_t dvd_cell_filesize(const ifo_handle_t *vmg_ifo, const ifo_handle_t *vts_ifo, const uint16_t track_number, const uint8_t cell_number);
+uint64_t dvd_cell_filesize(const ifo_handle_t *vmg_ifo, const ifo_handle_t *vts_ifo, const uint16_t track_number, const uint8_t cell_number);
 
 /**
  * Track some possible indicators that the DVD is authored intentionally to break
