+++ dvd_vob.c
@@ -4,13 +4,13 @@
  * Functions used to get information about a DVD VOB
  */
 
-ssize_t dvd_vob_blocks(dvd_reader_t *dvdread_dvd, const uint16_t vts_number, const uint16_t vob_number) {
+uint64_t dvd_vob_blocks(dvd_reader_t *dvdread_dvd, const uint16_t vts_number, const uint16_t vob_number) {
 
 	if(!vob_number)
 		return 0;
 
-	ssize_t vob_blocks = 0;
-	ssize_t vob_filesize = 0;
+	uint64_t vob_blocks = 0;
+	uint64_t vob_filesize = 0;
 
 	vob_filesize = dvd_vob_filesize(dvdread_dvd, vts_number, vob_number);
 	vob_blocks = vob_filesize / DVD_VIDEO_LB_LEN;
@@ -19,16 +19,16 @@ ssize_t dvd_vob_blocks(dvd_reader_t *dvdread_dvd, const uint16_t vts_number, con
 
 }
 
-ssize_t dvd_vob_filesize(dvd_reader_t *dvdread_dvd, const uint16_t vts_number, const uint16_t vob_number) {
+uint64_t dvd_vob_filesize(dvd_reader_t *dvdread_dvd, const uint16_t vts_number, const uint16_t vob_number) {
 
 	if(!vob_number)
 		return 0;
 
 	dvd_stat_t dvdread_stat;
-	ssize_t vob_filesize = 0;
+	uint64_t vob_filesize = 0;
 
 	DVDFileStat(dvdread_dvd, vts_number, DVD_READ_TITLE_VOBS, &dvdread_stat);
-	vob_filesize = dvdread_stat.parts_size[vob_number - 1];
+	vob_filesize = (uint64_t)dvdread_stat.parts_size[vob_number - 1];
 
 	return vob_filesize;
